[Plugin] [1/9] Remove passing a 'method holder' object and properly use TypeScript for plugin. (#4210)
## Summary

Since we are using TypeScript for plugin now we should get rid of
passing the magic `t` object as a function argument and use proper
imports instead. Also we no longer import whole `BabelCore` and
`BabelTypes` but just the required objects. This is a direct follow up
to #4209

## Test plan

`Yarn jest`, these changes don't touch plugin's logic.