Add lacking transformation of useAnimatedScrollHandler function argument to plugin (#4355)
## Summary

`useAnimatedScrollHandler` can take a single function as an argument
instead of an object containing functions. Functionality for this
transformation was lacking in plugin and this PR adds it.

I duplicated current tests for `objectHooks` for both
`useAnimatedGestureHandler` and `useAnimatedScrollHandler` to have them
separated and added tests according to this PR.

## Test plan

Run tests and build Example app to see that `useAnimatedScrollHandler`
works properly now.