Persist current screen during fast refresh on web (#4343)
## Summary

Previously, WebExample used to return to home screen after each fast
refresh which was quite a poor developer experience. This PR adds
linking so that the current screen persists. Note that the missing back
button after fast refresh is a well-known issue in react-navigation for
web but fortunately the browser back button works just fine.


https://user-images.githubusercontent.com/20516055/230615121-4b6d96db-23ca-4540-974a-0f76f9869bc2.mp4

## Test plan

1. Launch WebExample
2. Open Bokeh example
3. Save BokehExample.tsx to trigger fast refresh
4. The app shouldn't go back to home screen