doc: added important tip when using android device (#4357)
<!-- Thanks for submitting a pull request! We appreciate you spending
the time to work on these changes. Please follow the template so that
the reviewers can easily understand what the code changes affect. -->

## Summary

<!-- Explain the motivation for this PR. Include "Fixes #<number>" if
applicable. -->
This pull request aims to improve the documentation on [Events
Fundamentals](https://docs.swmansion.com/react-native-reanimated/docs/fundamentals/events/)
by adding instructions for how to set up event handlers on Android
devices. Many developers, including myself, have had issues practicing
the example on Android devices, which can be confusing for those who are
not familiar with *react-native-gesture-handler*.

This addition would make it easier for developers to practice using
event handlers with *react-native-reanimated* on Android devices
directly from the docs and help avoid spending time researching why it
isn't working.

## Changes Made
This pull request adds a new section to the Events Fundamentals
documentation that explains how to enable event handlers on Android
devices. It includes simple instructions and code example to help
developers enable the event handlers.

## Test plan
<!-- Provide a minimal but complete code snippet that can be used to
test out this change along with instructions how to run it and a
description of the expected behavior. -->
As this is a documentation update, tests are not applicable. However,
reviewers can follow the instructions provided in the new section and
verify that the event handlers are working correctly on Android devices.

---------

Co-authored-by: Tomek Zawadzki <tomekzawadzki98@gmail.com>
Co-authored-by: Krzysztof Piaskowy <krzysztof.piaskowy@swmansion.com>