Convert to shareable directly from runOnUI (#4372)
## Summary

This PR adds additional call to convert worklet and arguments to
shareables directly from runOnUI such that we avoid doing the conversion
from the microtasks queue. The reason for that is that if the object to
be converted triggers an error (stack exceeded or cyclic object),
converting it from microtasks queue makes it difficult to narrow down
the actual object that triggered the exception.

Since make-shareable method keeps a cache of converted objects we can be
sure the conversion doesn't happen twice, but nevertheless we only do
that in DEV mode.


## Test plan

Run Example app.