Add Reanimated Beta Docs banner (#4490)
### Motivation

This PR adds a banner to the New Reanimated Docs on the existing docs.
This makes it easier for people to access and test the new site.
Additionaly it links to the discussion where people can leave feedback
regarding the new docs.

This PR removes the 2023 App.js Conf announcement bar. 

### Images 


![image](https://github.com/software-mansion/react-native-reanimated/assets/39658211/356b777c-1c51-4490-9a0e-6204d9965335)