Added getAnimatableRef function (#4533)
## Summary

Added an option for a component to specify which other component should
be animated when `createAnimatedComponent` is executed on it by adding
`getAnimatableRef` function (we might want to think of a better name).

This is useful for some 3rd party libraries which want to add support
for animating their component, but it's wrapped inside of another
component which is inaccessible to the user.

## Test plan

Included a simple example screen where inner `Rect` component is
animated when `createAnimatedComponent` is executed on the component
containing it.


https://github.com/software-mansion/react-native-reanimated/assets/31368152/96d2cd19-a4f9-48ac-bd23-f50746dc7e0b