fix: ios build error when using 'use_frameworks!' in Podfile (#4504)
when open 'use_frameworks!' in Podfile, build will fail because of
unable to find hermes related headers. We need to declare hermes
dependency explicit in podspec

<!-- Thanks for submitting a pull request! We appreciate you spending
the time to work on these changes. Please follow the template so that
the reviewers can easily understand what the code changes affect. -->

## Summary

<!-- Explain the motivation for this PR. Include "Fixes #<number>" if
applicable. -->

## Test plan

<!-- Provide a minimal but complete code snippet that can be used to
test out this change along with instructions how to run it and a
description of the expected behavior. -->

---------

Co-authored-by: Tomek Zawadzki <tomekzawadzki98@gmail.com>