Fix spring bug with extra type verification (#4512)
<!-- Thanks for submitting a pull request! We appreciate you spending
the time to work on these changes. Please follow the template so that
the reviewers can easily understand what the code changes affect. -->

## Summary

<!-- Explain the motivation for this PR. Include "Fixes #<number>" if
applicable. -->
There is a need to check if velocity from previous animation is
undefined.
(Hopefully such a bugs will be soon reported by typescript 😊)

Closes #4510

<table>
<tr><td>BEFORE</td><td>AFTER</td></tr>
<tr><td>


https://github.com/software-mansion/react-native-reanimated/assets/56199675/638f0fea-9671-42e6-9d05-2e6717cf3645

</td><td>


https://github.com/software-mansion/react-native-reanimated/assets/56199675/82d656f3-c342-4e8b-9d31-a4bbf4df12a8


</td></tr>
</table>
## Test plan
Tested on added to our bouncing box animation sequence from this issue
https://github.com/software-mansion/react-native-reanimated/issues/4510:

```diff

    .onFinalize(() => {
+      offset.value = withSequence(
+        withTiming(20, { duration: 75 }),
+        withTiming(-20, { duration: 75 }),
+        withTiming(20, { duration: 75 }),
+        withSpring(0, { damping: 50, stiffness: 500, mass: 2.8 })
+      );
    });
```


<!-- Provide a minimal but complete code snippet that can be used to
test out this change along with instructions how to run it and a
description of the expected behavior. -->

---------

Co-authored-by: Aleksandra Cynk <aleksandracynk@swmansion.com>