Improve plugin's unit tests quality (#4447)
<!-- Thanks for submitting a pull request! We appreciate you spending
the time to work on these changes. Please follow the template so that
the reviewers can easily understand what the code changes affect. -->

## Summary

This PR is a part of ongoing effort to improve Reanimated Babel plugin.

- Most of our tests based purely on snapshots, turning them into actual
regression tests only. This PR adds an actual check to each test on top
of existing snapshot test. On the other hand, some snapshot testing was
added to tests that weren't having them but could use them.
- Changed output log location from `jest tests fixture` to `/dev/null` -
to simplify snapshot update process - we never used that file anyway.
- Moved most of string checks to RegExp checks for clarity.

## Test plan

No other way than to run the tests and read the tests.

---------

Co-authored-by: Tomek Zawadzki <tomasz.zawadzki@swmansion.com>