Fix ChoreographerCallback double execution (#4579)
## Summary

This PR fixes an issue with long animations. On Android, it was possible
to execute ChoreographerCallback twice during the same frame (after
frame drop). In effect, existing animations were canceled due to this
if-condition in `requestAnimationFrame` method:
```js
nativeRequestAnimationFrame((timestamp) => {
  if (lastNativeAnimationFrameTimestamp >= timestamp) {
    // Make sure we only execute the callbacks once for a given frame
    return;
  }
...
```
To fix it, I moved this check to the native code. When double execution
happens I skip the execution of the JS callback, but I still need to
schedule another choreographer callback.

## How to test it?
```java
final long[] lastTimestamp = {0};
new GuardedFrameCallback(context) {
  @Override
  protected void doFrameGuarded(long frameTimeNanos) {
    if (lastTimestamp[0] == frameTimeNanos) {
      Log.w("test");
    }
    onAnimationFrame(frameTimeNanos);
    lastTimestamp[0] = frameTimeNanos;
  }
}
```
### Before
- Open SVGExample
- Set a breakpoint in Android Studio at line `Log.w("test");`
- Wait some time until the breakpoint hit
- Pass breakpoint
- Animation will **stop** ❌

### After
- Open SVGExample
- Set a breakpoint in Android Studio at line `Log.w("test");`
- Wait some time until the breakpoint hit
- Pass breakpoint
- Animation will **keep going**  ✅