Fix R class reference ReanimatedKeyboardEventListener (#4571)
## Summary
In order to upgrade to Android Gradle Plugin (AGP) 8.0, we need to
update app build files to accommodate five important build behavior
changes. This PR fixes R class reference in
`ReanimatedKeyboardEventListener` which will enable us to enable
`nonTransitiveRClass` as a followup. This behaviour is enforced in AGP
8.0+.

As a side bonus of enabling non-transitive R class, we will get app size
decrease on Android, as R classes are quite big.

You can read more about it here
https://medium.com/androiddevelopers/5-ways-to-prepare-your-app-build-for-android-studio-flamingo-release-da34616bb946

Fixes #4463 

## Test
Please test the flows that involve this method call to make sure it
doesn't crash and work as intended.

Co-authored-by: Krzysztof Piaskowy <krzysztof.piaskowy@swmansion.com>