Fix sensors on android web (#4499)
<!-- Thanks for submitting a pull request! We appreciate you spending
the time to work on these changes. Please follow the template so that
the reviewers can easily understand what the code changes affect. -->

## Summary

### Problem

Android devices use a right-handed coordinate system, where the positive
x-axis points to the right, the positive y-axis points up, and the
positive z-axis points out of the screen.
However iOS uses a left-handed coordinate system, where the positive
x-axis points to the right, the positive y-axis points down, and the
positive z-axis points into the screen.

This is already fixed on android native code, but the problem still
exists on android web browser (react native web).

### Fix

- Replace quaternion z, y and negate z  
- Do the same thing on Android native

## Test plan

- Web Example App