Fix babel plugins' location (#4561)
## Summary
When applying custom folders or custom configuration for metro bundler
it gives an error saying "Cannot find module
"@babel/plugin-transform-shorthands-properties". Also its a good
practice to rely on node to resolve the location of the package instead
of just provide the module name. It might also be caused by
`@babel/register` which in some cases cannot find the packages even
though they are in the same node_modules.

Example: run metro with a custom folder and adding `extraNodeModules` to
point to another folder outside of the current project.

## Test plan

Run metro with custom config and point react-native-reanimated as
extraNodeModules to a specific folder outside of main project.

---------

Co-authored-by: Tomasz Żelawski <40713406+tjzel@users.noreply.github.com>