Unify platform checks (#4594)
<!-- Thanks for submitting a pull request! We appreciate you spending
the time to work on these changes. Please follow the template so that
the reviewers can easily understand what the code changes affect. -->

## Summary

This PR replaces direct platform checks such as `Platform.OS ===
'android'` or `process.env.JEST_WORKER_ID` with functions from
`PlatformChecker.ts` (or captured constants when used inside worklets).

## Test plan

<!-- Provide a minimal but complete code snippet that can be used to
test out this change along with instructions how to run it and a
description of the expected behavior. -->