Move `scheduleOnJS` to a separate method (#4621)
<!-- Thanks for submitting a pull request! We appreciate you spending
the time to work on these changes. Please follow the template so that
the reviewers can easily understand what the code changes affect. -->

## Summary

This PR moves the inline implementation of `scheduleOnJS` from C++
lambda body to a separate method of `NativeReanimatedModule` so that it
can be re-used for different purposes 😉

## Test plan

WorkletExample.tsx &rarr; runOnUI + runOnJS demo